rename querydefinitions -> querydefinition
