+++ aws-logs-querydefinition/.gitignore
