+++ aws-logs-querydefinition/lombok.config
