+++ aws-logs-querydefinition/resource-role.yaml
