+++ aws-logs-querydefinition/src/main/java/software/amazon/logs/querydefinition/CallbackContext.java
@@ -1,0 +1,11 @@
+package software.amazon.logs.querydefinition;
+
+import software.amazon.cloudformation.proxy.StdCallbackContext;
+
+@lombok.Getter
+@lombok.Setter
+@lombok.ToString
+@lombok.EqualsAndHashCode(callSuper = true)
+public class CallbackContext extends StdCallbackContext {
+    private String QueryDefinitionId;
+}
