+++ aws-logs-querydefinition/src/main/java/software/amazon/logs/querydefinition/Configuration.java
@@ -1,0 +1,8 @@
+package software.amazon.logs.querydefinition;
+
+class Configuration extends BaseConfiguration {
+
+    public Configuration() {
+        super("aws-logs-querydefinition.json");
+    }
+}
