+++ aws-logs-querydefinition/src/resources/log4j2.xml
