+++ aws-logs-querydefinition/template.yml
@@ -1,0 +1,23 @@
+AWSTemplateFormatVersion: "2010-09-09"
+Transform: AWS::Serverless-2016-10-31
+Description: AWS SAM template for the AWS::Logs::QueryDefinition resource type
+
+Globals:
+  Function:
+    Timeout: 180  # docker start-up times can be long for SAM CLI
+    MemorySize: 256
+
+Resources:
+  TypeFunction:
+    Type: AWS::Serverless::Function
+    Properties:
+      Handler: software.amazon.logs.querydefinition.HandlerWrapper::handleRequest
+      Runtime: java8
+      CodeUri: ./target/aws-logs-querydefinition-handler-1.0-SNAPSHOT.jar
+
+  TestEntrypoint:
+    Type: AWS::Serverless::Function
+    Properties:
+      Handler: software.amazon.logs.querydefinition.HandlerWrapper::testEntrypoint
+      Runtime: java8
+      CodeUri: ./target/aws-logs-querydefinition-handler-1.0-SNAPSHOT.jar
