--- aws-logs-querydefinitions/.rpdk-config
@@ -1,25 +1,0 @@
-{
-    "artifact_type": "RESOURCE",
-    "typeName": "AWS::Logs::QueryDefinitions",
-    "language": "java",
-    "runtime": "java8",
-    "entrypoint": "software.amazon.logs.querydefinitions.HandlerWrapper::handleRequest",
-    "testEntrypoint": "software.amazon.logs.querydefinitions.HandlerWrapper::testEntrypoint",
-    "settings": {
-        "version": false,
-        "subparser_name": null,
-        "verbose": 0,
-        "force": false,
-        "type_name": null,
-        "artifact_type": null,
-        "namespace": [
-            "software",
-            "amazon",
-            "logs",
-            "querydefinitions"
-        ],
-        "codegen_template_path": "default",
-        "protocolVersion": "2.0.0"
-    },
-    "executableEntrypoint": "software.amazon.logs.querydefinitions.HandlerWrapperExecutable"
-}
