--- aws-logs-querydefinitions/src/main/java/software/amazon/logs/querydefinitions/CallbackContext.java
@@ -1,11 +1,0 @@
-package software.amazon.logs.querydefinitions;
-
-import software.amazon.cloudformation.proxy.StdCallbackContext;
-
-@lombok.Getter
-@lombok.Setter
-@lombok.ToString
-@lombok.EqualsAndHashCode(callSuper = true)
-public class CallbackContext extends StdCallbackContext {
-    private String QueryDefinitionId;
-}
