--- aws-logs-querydefinitions/src/main/java/software/amazon/logs/querydefinitions/Configuration.java
@@ -1,8 +1,0 @@
-package software.amazon.logs.querydefinitions;
-
-class Configuration extends BaseConfiguration {
-
-    public Configuration() {
-        super("aws-logs-querydefinitions.json");
-    }
-}
